feat: 减少第三方依赖
