+++ src/controller/openai/config.ts
@@ -14,7 +14,7 @@ export const CONFIG: OpenAiConfig = {
   OPEN_AI_FREE_USAGE: 'https://api.openai.com/dashboard/billing/credit_grants',
   // OpenAI API 额外参数 TODO
   OPEN_AI_API_EXTRA_PARAMS: {},
-  // OpenAI API 请求超时毫秒
+  // OpenAI API 请求超时，毫秒
   OPEN_AI_API_TIMEOUT_MS: 30000,
   // 单次请求 OpenAI 最大 token 数
   MAX_CHAT_TOKEN_NUM: 4000,
@@ -22,7 +22,6 @@ export const CONFIG: OpenAiConfig = {
   MIN_CHAT_RESPONSE_TOKEN_NUM: 500,
   // 串聊最大历史记录长度
   MAX_HISTORY_LENGTH: 20,
-  // 全局默认初始化消息，不能在一开始初始化，会为 0，cloudflare 的限制
-  SYSTEM_INIT_MESSAGE: () => `You are ChatGPT, a large language model trained by OpenAI. Answer as concisely as possible. Knowledge cutoff: 2021-09-01. Current date: ${new Date().toISOString().split('T')[0]
-  }`,
+  // 全局默认初始化消息，不能使用 Date.now() 获取当前时间，实际会为 0，cloudflare 的限制
+  SYSTEM_INIT_MESSAGE: `You are ChatGPT, a large language model trained by OpenAI. Answer as concisely as possible. Knowledge cutoff: 2021-09-01. Current is 2023`,
 }
